Pretty printing, especially for ForAll, needs access to inferred types.
So I guess we need an environment-ish around in order to pretty print.
This doesn't fully thread that through everywhere, it just reintroduces
the notion of `constants` which seems like the "real" environment builds
up with what it has already type checked (so we do that too now).

I'm surprised this translates (because the `constants` parameter default
is a dict but in Real Life it's an r_dict). But we can fiddle with that.

This at least passes one of the non-trivial pretty printing tests for
ForAll.